let's implement a feature that can be useful in the greek Lexicon tool as well as the Hebrew lexicon tool.
here is how the feature should work.

implement a LookupHistory feature component that stores (and later displays) every word successfully looked up in the lexicon. This will help users easily lookup previously seen words without the need to type them in again through the input  field. the feature should store words by namespace just as the vocabulary browser feature. When the feature component is in view, it should fetch list of namespaces and populate the namespace list so users can select which namespace history they want to pick words from.

here is the how the structure of the lookup list data from the backend:

[
  {
    "id": 1,
    "created_at": "2025-05-27 13:15:48.390027+00",
    "language": "greek",
    "namespace": "test namespace",
    "word": "λόγος",
    "lemma": "λόγος",
    "frequency": 1,
    "updated_at": "2025-05-27 13:15:48.390027"
  }
]